requirements.txt
flask
psycopg2-binary
python-dotenv
flask-login
gunicorn
flask-limiter
flask-wtf
sib-api-v3-sdk
stripe